feat(answer): add delete useCase

+++ src/domain/forum/application/use-cases/delete-answer.ts
@@ -2,6 +2,7 @@ import { AnswersRepository } from "../repositories/answers-repository";
 
 interface DeleteAnswerUseCaseRequest {
   id: string;
+  authorId: string;
 }
 
 interface DeleteAnswerUseCaseResponse {}
@@ -15,10 +16,15 @@ export class DeleteAnswerUseCase {
 
   async execute({
     id,
+    authorId,
   }: DeleteAnswerUseCaseRequest): Promise<DeleteAnswerUseCaseResponse> {
     const answer = await this.answersRepository.getById(id);
 
     if (!answer) throw new Error("answer not found");
+
+    if (answer.authorId.toString() != authorId) {
+      throw new Error("not allowed");
+    }
 
     await this.answersRepository.delete(answer);
